WIP: moving Kwik store items.
