+++ phy/io/kwik/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+# flake8: noqa
+
+"""Kwik format."""
+
+from .model import KwikModel
