+++ phy/utils/__init__.py
@@ -1,9 +1,7 @@
 # -*- coding: utf-8 -*-
 # flake8: noqa
 
-"""Utilities.
-
-"""
+"""Utilities."""
 
 from .logging import debug, info, warn, register, unregister, set_level
 from ._types import _is_array_like, _as_array, _as_tuple, _as_list, Bunch
@@ -16,3 +14,4 @@ from .array import (_unique,
 from .event import EventEmitter, ProgressReporter
 from .dock import DockWindow, start_qt_app, run_qt_app, qt_app, enable_qt
 from .datasets import download_file, download_test_data
+from .selector import Selector
