+++ phy/utils/tests/test_selector.py
@@ -9,8 +9,8 @@
 import numpy as np
 from numpy.testing import assert_array_equal as ae
 
-from ....io.mock import artificial_spike_clusters
-from ....utils import _spikes_in_clusters
+from ...io.mock import artificial_spike_clusters
+from ..array import _spikes_in_clusters
 from ..selector import Selector
